change dtype of charges in ChargesContainer from uint16 to int32

+++ src/nectarchain/makers/component/tests/test_charges_component.py
@@ -149,11 +149,11 @@ class TestChargesComponent(BaseTestArrayDataComponent):
 
         assert np.all(
             instance.charges_hg(event.trigger.event_type)
-            == np.uint16(instance._charges_hg[f"{name}"])
+            == np.int32(instance._charges_hg[f"{name}"])
         )
         assert np.all(
             instance.charges_lg(event.trigger.event_type)
-            == np.uint16(instance._charges_lg[f"{name}"])
+            == np.int32(instance._charges_lg[f"{name}"])
         )
         assert np.all(
             instance.peak_hg(event.trigger.event_type)
